Take out deprecated methods
