+++ framework/src/play-java-jdbc/src/main/java/play/db/ConnectionPool.java
@@ -25,21 +25,6 @@ public interface ConnectionPool {
     DataSource create(String name, Config configuration, Environment environment);
 
     /**
-     * Create a data source with the given configuration.
-     *
-     * @param name the database name
-     * @param configuration the data source configuration
-     * @param environment the database environment
-     * @return a data source backed by a connection pool
-     *
-     * @deprecated Use create(String, Config, Environment
-     */
-    @Deprecated
-    default DataSource create(String name, Configuration configuration, Environment environment) {
-        return create(name, configuration.underlying(), environment);
-    }
-
-    /**
      * Close the given data source.
      *
      * @param dataSource the data source to close
