+++ framework/src/play-java-jdbc/src/main/java/play/db/DBModule.java
@@ -3,21 +3,17 @@
  */
 package play.db;
 
-import java.util.Set;
-import javax.inject.Inject;
-import javax.inject.Provider;
-
-import scala.collection.Seq;
-
+import com.google.common.collect.ImmutableList;
 import play.api.Configuration;
 import play.api.Environment;
 import play.api.inject.Binding;
 import play.api.inject.Module;
-import play.db.NamedDatabase;
-import play.db.NamedDatabaseImpl;
 import play.libs.Scala;
+import scala.collection.Seq;
 
-import com.google.common.collect.ImmutableList;
+import javax.inject.Inject;
+import javax.inject.Provider;
+import java.util.Set;
 
 /**
  * Injection module with default DB components.
