+++ framework/src/play-java-jdbc/src/main/java/play/db/DefaultDatabase.java
@@ -10,7 +10,8 @@ import javax.sql.DataSource;
 import com.typesafe.config.Config;
 
 import com.typesafe.config.ConfigFactory;
-import play.Configuration;
+import scala.runtime.AbstractFunction1;
+import scala.runtime.BoxedUnit;
 
 /**
  * Default delegating implementation of the database API.
@@ -33,19 +34,6 @@ public class DefaultDatabase implements Database {
         this(new play.api.db.PooledDatabase(name, new play.api.Configuration(
                 configuration.withFallback(ConfigFactory.defaultReference().getConfig("play.db.prototype"))
         )));
-    }
-
-    /**
-     * Create a default BoneCP-backed database.
-     *
-     * @param name name for the db's underlying datasource
-     * @param configuration the database's configuration
-     *
-     * @deprecated use the version that accepts Config
-     */
-    @Deprecated
-    public DefaultDatabase(String name, Configuration configuration) {
-        this(name, configuration.underlying());
     }
 
     /**
@@ -88,32 +76,32 @@ public class DefaultDatabase implements Database {
 
     @Override
     public void withConnection(ConnectionRunnable block) {
-        db.withConnection(DB.connectionFunction(block));
+        db.withConnection(connectionFunction(block));
     }
 
     @Override
     public <A> A withConnection(ConnectionCallable<A> block) {
-        return db.withConnection(DB.connectionFunction(block));
+        return db.withConnection(connectionFunction(block));
     }
 
     @Override
     public void withConnection(boolean autocommit, ConnectionRunnable block) {
-        db.withConnection(autocommit, DB.connectionFunction(block));
+        db.withConnection(autocommit, connectionFunction(block));
     }
 
     @Override
     public <A> A withConnection(boolean autocommit, ConnectionCallable<A> block) {
-        return db.withConnection(autocommit, DB.connectionFunction(block));
+        return db.withConnection(autocommit, connectionFunction(block));
     }
 
     @Override
     public void withTransaction(ConnectionRunnable block) {
-        db.withTransaction(DB.connectionFunction(block));
+        db.withTransaction(connectionFunction(block));
     }
 
     @Override
     public <A> A withTransaction(ConnectionCallable<A> block) {
-        return db.withTransaction(DB.connectionFunction(block));
+        return db.withTransaction(connectionFunction(block));
     }
 
     @Override
@@ -125,4 +113,44 @@ public class DefaultDatabase implements Database {
     public play.api.db.Database toScala() {
         return db;
     }
+
+
+    /**
+     * Create a Scala function wrapper for ConnectionRunnable.
+     *
+     * @param block a Java functional interface instance to wrap
+     * @return a scala function that wraps the given block
+     */
+    AbstractFunction1<Connection, BoxedUnit> connectionFunction(final ConnectionRunnable block) {
+        return new AbstractFunction1<Connection, BoxedUnit>() {
+            public BoxedUnit apply(Connection connection) {
+                try {
+                    block.run(connection);
+                    return BoxedUnit.UNIT;
+                } catch (java.sql.SQLException e) {
+                    throw new RuntimeException("Connection runnable failed", e);
+                }
+            }
+        };
+    }
+
+    /**
+     * Create a Scala function wrapper for ConnectionCallable.
+     *
+     * @param block a Java functional interface instance to wrap
+     * @param <A> the provided block's return type
+     * @return a scala function wrapping the given block
+     */
+    <A> AbstractFunction1<Connection, A> connectionFunction(final ConnectionCallable<A> block) {
+        return new AbstractFunction1<Connection, A>() {
+            public A apply(Connection connection) {
+                try {
+                    return block.call(connection);
+                } catch (java.sql.SQLException e) {
+                    throw new RuntimeException("Connection callable failed", e);
+                }
+            }
+        };
+    }
+
 }
